requirements.txt
urllib3
Pillow
